Fix warning in Spells.cpp: change variable to size_t.

+++ src/game/Spells.cpp
@@ -3877,8 +3877,7 @@ bool ARX_SPELLS_Launch( const long& typ, const long& source, const long& flagss,
 						}
 					}
 
-				for (long n=0;n<MAX_SPELLS;n++)
-				{
+				for(size_t n = 0; n < MAX_SPELLS; n++) {
 					if (!spells[n].exist) continue;
 
 					switch (spells[n].type)
